Reorganized component folder
