+++ src/components/bottom/BottomPanel.tsx
@@ -1,8 +1,8 @@
 import React, { Component } from 'react';
 import { connect } from 'react-redux';
-import { handleClose, deleteProp, addProp } from '../actions/actionCreators';
+import { handleClose, deleteProp, addProp } from '../../actions/actionCreators';
 import BottomTabs from './BottomTabs';
-import { PropsInt, PropInt } from '../utils/Interfaces';
+import { PropsInt, PropInt } from '../../intefaces/Interfaces';
 
 const IPC = require('electron').ipcRenderer;
