+++ src/components/bottom/Props.tsx
@@ -7,9 +7,9 @@ import TextField from '@material-ui/core/TextField';
 import Button from '@material-ui/core/Button';
 import Select from '@material-ui/core/Select';
 import InputLabel from '@material-ui/core/InputLabel';
-import { addProp, deleteProp } from '../actions/actionCreators';
+import { addProp, deleteProp } from '../../actions/actionCreators';
 import DataTable from './DataTable';
-import { PropInt, PropsInt } from '../utils/Interfaces';
+import { PropInt, PropsInt } from '../../intefaces/Interfaces';
 
 interface PropsPropsInt extends PropsInt {
   classes: any;
