+++ src/components/main/TransformerComponent.tsx
@@ -1,6 +1,6 @@
 import React, { Component } from 'react';
 import { Transformer } from 'react-konva';
-import { ChildInt } from '../utils/Interfaces';
+import { ChildInt } from '../../intefaces/Interfaces';
 import Konva from 'konva';
 
 interface TCompPropsInt {
