+++ src/helperFunctions/isEmpty.ts
