+++ src/localStorage.ts
@@ -1,5 +1,5 @@
 import localforage from 'localforage';
-import { ApplicationStateInt } from './utils/Interfaces';
+import { ApplicationStateInt } from './intefaces/Interfaces';
 
 export const saveState = (state: ApplicationStateInt) =>
   localforage.setItem('state-v1.0.1', state);
